[Compose] - feature/AA-1749-add-jetpack-compose
- Review clean and fixes
