--- library-compose/src/main/res/font/gt_america.xml
@@ -1,25 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<font-family xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:app="http://schemas.android.com/apk/res-auto"
-    xmlns:tools="http://schemas.android.com/tools">
-
-    <font
-        android:font="@font/font_raw_gt_america_regular"
-        android:fontStyle="normal"
-        android:fontWeight="400"
-
-        app:font="@font/font_raw_gt_america_regular"
-        app:fontStyle="normal"
-        app:fontWeight="400"
-        tools:targetApi="o" />
-
-    <font
-        android:font="@font/font_raw_gt_america_regular_italic"
-        android:fontStyle="italic"
-        android:fontWeight="400"
-
-        app:font="@font/font_raw_gt_america_regular_italic"
-        app:fontStyle="italic"
-        app:fontWeight="400"
-        tools:targetApi="o" />
-</font-family>
