--- library-compose/src/main/res/font/gt_america_bold.xml
@@ -1,22 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<font-family xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:app="http://schemas.android.com/apk/res-auto">
-
-    <font
-        android:font="@font/font_raw_gt_america_medium"
-        android:fontStyle="normal"
-        android:fontWeight="700"
-
-        app:font="@font/font_raw_gt_america_medium"
-        app:fontStyle="normal"
-        app:fontWeight="700" />
-
-    <font
-        android:font="@font/font_raw_gt_america_medium_italic"
-        android:fontStyle="italic"
-        android:fontWeight="700"
-
-        app:font="@font/font_raw_gt_america_medium_italic"
-        app:fontStyle="italic"
-        app:fontWeight="700" />
-</font-family>
